requirements.txt
flask==3.0.2
flask-cors==4.0.0
werkzeug==3.0.1
pyjwt==2.8.0 